Add client-side and server-side validation
This even caught a bug: timestamps were sent as floating numbers!

+++ client/src/client.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
 import * as api from "./api";
+import * as validation from "./validation";
 
 const URL = "http://localhost:3010";
 
@@ -7,6 +8,12 @@ export async function registerUser(
   headers: api.AuthOptional,
   request: api.RegisterUserRequest,
 ): Promise<api.RegisterUserResponse> {
+  if (!validation.validate_AuthOptional(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_RegisterUserRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/users/register`;
   const response = await axios({
     url,
@@ -14,30 +21,51 @@ export async function registerUser(
     data: request,
     headers,
   });
+  if (!validation.validate_RegisterUserResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
 export async function loginUser(
   request: api.LoginUserRequest,
 ): Promise<api.LoginUserResponse> {
+  if (!validation.validate_LoginUserRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/users/login`;
   const response = await axios({
     url,
     method: "POST",
     data: request,
   });
+  if (!validation.validate_LoginUserResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
 export async function checkAuth(
   headers: api.AuthRequired,
 ): Promise<api.LoginUserResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
   const url = `${URL}/users/auth`;
   const response = await axios({
     url,
     method: "POST",
     headers,
   });
+  if (!validation.validate_LoginUserResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -46,6 +74,12 @@ export async function updateUser(
   id: string,
   request: api.UpdateUserRequest,
 ): Promise<api.UpdateUserResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_UpdateUserRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/users/${id}`;
   const response = await axios({
     url,
@@ -53,6 +87,11 @@ export async function updateUser(
     data: request,
     headers,
   });
+  if (!validation.validate_UpdateUserResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -61,6 +100,12 @@ export async function deleteUser(
   id: string,
   request: api.DeleteUserRequest,
 ): Promise<api.DeleteUserResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_DeleteUserRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/users/${id}`;
   const response = await axios({
     url,
@@ -68,6 +113,11 @@ export async function deleteUser(
     data: request,
     headers,
   });
+  if (!validation.validate_DeleteUserResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -75,6 +125,12 @@ export async function listUsers(
   headers: api.AuthRequired,
   request: api.ListUsersRequest,
 ): Promise<api.ListUsersResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_ListUsersRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/users/list`;
   const response = await axios({
     url,
@@ -82,6 +138,11 @@ export async function listUsers(
     data: request,
     headers,
   });
+  if (!validation.validate_ListUsersResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -89,6 +150,12 @@ export async function createApartment(
   headers: api.AuthRequired,
   request: api.CreateApartmentRequest,
 ): Promise<api.CreateApartmentResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_CreateApartmentRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/apartments/create`;
   const response = await axios({
     url,
@@ -96,6 +163,11 @@ export async function createApartment(
     data: request,
     headers,
   });
+  if (!validation.validate_CreateApartmentResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -104,6 +176,12 @@ export async function updateApartment(
   id: string,
   request: api.UpdateApartmentRequest,
 ): Promise<api.UpdateApartmentResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_UpdateApartmentRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/apartments/${id}`;
   const response = await axios({
     url,
@@ -111,6 +189,11 @@ export async function updateApartment(
     data: request,
     headers,
   });
+  if (!validation.validate_UpdateApartmentResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -118,12 +201,20 @@ export async function deleteApartment(
   headers: api.AuthRequired,
   id: string,
 ): Promise<api.DeleteApartmentResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
   const url = `${URL}/apartments/${id}`;
   const response = await axios({
     url,
     method: "DELETE",
     headers,
   });
+  if (!validation.validate_DeleteApartmentResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
 
@@ -131,6 +222,12 @@ export async function listApartments(
   headers: api.AuthRequired,
   request: api.ListApartmentsRequest,
 ): Promise<api.ListApartmentsResponse> {
+  if (!validation.validate_AuthRequired(headers)) {
+    throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+  }
+  if (!validation.validate_ListApartmentsRequest(request)) {
+    throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+  }
   const url = `${URL}/apartments/list`;
   const response = await axios({
     url,
@@ -138,5 +235,10 @@ export async function listApartments(
     data: request,
     headers,
   });
+  if (!validation.validate_ListApartmentsResponse(response.data)) {
+    throw new Error(
+      `Invalid response: ${JSON.stringify(response.data, null, 2)}`,
+    );
+  }
   return response.data;
 }
