+++ server/src/server.ts
@@ -1,11 +1,11 @@
 import "module-alias/register";
 
 import { initDatabase } from "@/db";
-import { checkAuth } from "@/endpoints/checkAuth";
 import bodyParser from "body-parser";
 import cors from "cors";
 import express from "express";
 import * as api from "./api";
+import { checkAuth } from "./endpoints/checkAuth";
 import { createApartment } from "./endpoints/createApartment";
 import { deleteApartment } from "./endpoints/deleteApartment";
 import { deleteUser } from "./endpoints/deleteUser";
@@ -15,6 +15,7 @@ import { loginUser } from "./endpoints/loginUser";
 import { registerUser } from "./endpoints/registerUser";
 import { updateApartment } from "./endpoints/updateApartment";
 import { updateUser } from "./endpoints/updateUser";
+import * as validation from "./validation";
 
 const PORT = 3010;
 
@@ -41,11 +42,20 @@ app.post("/users/register", async (req, res, next) => {
     const headers: api.AuthOptional = {
       Authorization: req.header("Authorization"),
     };
+    if (!validation.validate_AuthOptional(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const request: api.RegisterUserRequest = req.body;
+    if (!validation.validate_RegisterUserRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.RegisterUserResponse = await registerUser(
       headers,
       request,
     );
+    if (!validation.validate_RegisterUserResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -55,7 +65,13 @@ app.post("/users/register", async (req, res, next) => {
 app.post("/users/login", async (req, res, next) => {
   try {
     const request: api.LoginUserRequest = req.body;
+    if (!validation.validate_LoginUserRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.LoginUserResponse = await loginUser(request);
+    if (!validation.validate_LoginUserResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -65,9 +81,15 @@ app.post("/users/login", async (req, res, next) => {
 app.post("/users/auth", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const response: api.LoginUserResponse = await checkAuth(headers);
+    if (!validation.validate_LoginUserResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -77,15 +99,24 @@ app.post("/users/auth", async (req, res, next) => {
 app.put("/users/:id", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const id = req.params.id;
     const request: api.UpdateUserRequest = req.body;
+    if (!validation.validate_UpdateUserRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.UpdateUserResponse = await updateUser(
       headers,
       id,
       request,
     );
+    if (!validation.validate_UpdateUserResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -95,15 +126,24 @@ app.put("/users/:id", async (req, res, next) => {
 app.delete("/users/:id", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const id = req.params.id;
     const request: api.DeleteUserRequest = req.body;
+    if (!validation.validate_DeleteUserRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.DeleteUserResponse = await deleteUser(
       headers,
       id,
       request,
     );
+    if (!validation.validate_DeleteUserResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -113,10 +153,19 @@ app.delete("/users/:id", async (req, res, next) => {
 app.post("/users/list", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const request: api.ListUsersRequest = req.body;
+    if (!validation.validate_ListUsersRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.ListUsersResponse = await listUsers(headers, request);
+    if (!validation.validate_ListUsersResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -126,13 +175,22 @@ app.post("/users/list", async (req, res, next) => {
 app.post("/apartments/create", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const request: api.CreateApartmentRequest = req.body;
+    if (!validation.validate_CreateApartmentRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.CreateApartmentResponse = await createApartment(
       headers,
       request,
     );
+    if (!validation.validate_CreateApartmentResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -142,15 +200,24 @@ app.post("/apartments/create", async (req, res, next) => {
 app.put("/apartments/:id", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const id = req.params.id;
     const request: api.UpdateApartmentRequest = req.body;
+    if (!validation.validate_UpdateApartmentRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.UpdateApartmentResponse = await updateApartment(
       headers,
       id,
       request,
     );
+    if (!validation.validate_UpdateApartmentResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -160,13 +227,19 @@ app.put("/apartments/:id", async (req, res, next) => {
 app.delete("/apartments/:id", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const id = req.params.id;
     const response: api.DeleteApartmentResponse = await deleteApartment(
       headers,
       id,
     );
+    if (!validation.validate_DeleteApartmentResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
@@ -176,13 +249,22 @@ app.delete("/apartments/:id", async (req, res, next) => {
 app.post("/apartments/list", async (req, res, next) => {
   try {
     const headers: api.AuthRequired = {
-      Authorization: req.header("Authorization") || "",
+      Authorization: req.header("Authorization")!,
     };
+    if (!validation.validate_AuthRequired(headers)) {
+      throw new Error(`Invalid headers: ${JSON.stringify(headers, null, 2)}`);
+    }
     const request: api.ListApartmentsRequest = req.body;
+    if (!validation.validate_ListApartmentsRequest(request)) {
+      throw new Error(`Invalid request: ${JSON.stringify(request, null, 2)}`);
+    }
     const response: api.ListApartmentsResponse = await listApartments(
       headers,
       request,
     );
+    if (!validation.validate_ListApartmentsResponse(response)) {
+      throw new Error(`Invalid response: ${JSON.stringify(response, null, 2)}`);
+    }
     res.json(response);
   } catch (err) {
     next(err);
